PAYARA-986 Add -configuration suffix to the set commands of both request tracing and notification service.
