+++ nucleus/payara-modules/requesttracing-core/src/main/java/fish/payara/nucleus/requesttracing/admin/SetRequestTracingConfiguration.java
@@ -1,16 +1,27 @@
 /*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
+
+ DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
+
+ Copyright (c) 2016 C2B2 Consulting Limited. All rights reserved.
+
+ The contents of this file are subject to the terms of the Common Development
+ and Distribution License("CDDL") (collectively, the "License").  You
+ may not use this file except in compliance with the License.  You can
+ obtain a copy of the License at
+ https://glassfish.dev.java.net/public/CDDL+GPL_1_1.html
+ or packager/legal/LICENSE.txt.  See the License for the specific
+ language governing permissions and limitations under the License.
+
+ When distributing the software, include this License Header Notice in each
+ file and include the License file at packager/legal/LICENSE.txt.
  */
-package fish.payara.nucleus.notification.admin;
+package fish.payara.nucleus.requesttracing.admin;
 
 import com.sun.enterprise.config.serverbeans.Domain;
 import com.sun.enterprise.util.SystemPropertyConstants;
-import fish.payara.nucleus.notification.domain.execoptions.NotifierConfigurationExecutionOptionsFactory;
 import java.util.Properties;
+import java.util.logging.Logger;
 import javax.inject.Inject;
-import static javax.management.Query.value;
 import org.glassfish.api.ActionReport;
 import org.glassfish.api.I18n;
 import org.glassfish.api.Param;
@@ -30,23 +41,26 @@ import org.glassfish.hk2.api.ServiceLocator;
 import org.jvnet.hk2.annotations.Service;
 
 /**
- * Admin command to set notification services configuration
+ * Admin command to set Request Tracing services configuration
  *
  * @author Susan Rai
  */
-@ExecuteOn(value = {RuntimeType.DAS})
+@ExecuteOn({RuntimeType.DAS})
 @TargetType(value = {CommandTarget.DAS, CommandTarget.STANDALONE_INSTANCE, CommandTarget.CLUSTER, CommandTarget.CLUSTERED_INSTANCE, CommandTarget.CONFIG})
-@Service(name = "set-notification")
+@Service(name = "set-requesttracing-configuration")
 @CommandLock(CommandLock.LockType.NONE)
 @PerLookup
-@I18n("set.notification")
+@I18n("set.requesttracing.configuration")
 @RestEndpoints({
     @RestEndpoint(configBean = Domain.class,
             opType = RestEndpoint.OpType.POST,
-            path = "set-notification",
-            description = "Set notification Services Configuration")
+            path = "set-requesttracing-configuration",
+            description = "Set Request Tracing Services Configuration")
 })
-public class SetNotification implements AdminCommand {
+public class SetRequestTracingConfiguration implements AdminCommand {
+
+    @Inject
+    protected Logger logger;
 
     @Param(name = "target", optional = true, defaultValue = SystemPropertyConstants.DAS_SERVER_NAME)
     String target;
@@ -55,10 +69,16 @@ public class SetNotification implements AdminCommand {
     private Boolean enabled;
 
     @Param(name = "dynamic", optional = true, defaultValue = "false")
-    protected Boolean dynamic;
+    private Boolean dynamic;
+
+    @Param(name = "thresholdUnit", optional = true, defaultValue = "SECONDS")
+    private String unit;
+
+    @Param(name = "thresholdValue", optional = true, defaultValue = "30")
+    private String value;
 
     @Param(name = "notifierDynamic", optional = true, defaultValue = "false")
-    protected Boolean notifierDynamic;
+    private Boolean notifierDynamic;
 
     @Param(name = "notifierEnabled", optional = false)
     private Boolean notifierEnabled;
@@ -81,6 +101,10 @@ public class SetNotification implements AdminCommand {
             actionReport.setExtraProperties(extraProperties);
         }
 
+        if (!validate(actionReport)) {
+            return;
+        }
+
         if (dynamic || enabled) {
             if (dynamic) {
                 notifierDynamic = true;
@@ -92,28 +116,30 @@ public class SetNotification implements AdminCommand {
             } else {
                 notifierEnabled = false;
             }
-            enableNotificationConfigureOnTarget(actionReport, theContext, enabled);
+            enableRequestTracingConfigureOnTarget(actionReport, theContext, enabled);
         }
 
         if (notifierDynamic || notifierEnabled) {
-            enableNotificationNotifierConfigurerOnTarget(actionReport, theContext, notifierEnabled);
+            enableRequestTracingNotifierConfigurerOnTarget(actionReport, theContext, notifierEnabled);
         }
     }
 
-    private void enableNotificationConfigureOnTarget(ActionReport actionReport, AdminCommandContext context, Boolean enabled) {
+    private void enableRequestTracingConfigureOnTarget(ActionReport actionReport, AdminCommandContext context, Boolean enabled) {
         CommandRunner runner = serviceLocator.getService(CommandRunner.class);
         ActionReport subReport = context.getActionReport().addSubActionsReport();
 
         if (target.equals("server-config")) {
-            inv = runner.getCommandInvocation("notification-configure-das", subReport, context.getSubject());
+            inv = runner.getCommandInvocation("requesttracing-configure-das", subReport, context.getSubject());
         } else {
-            inv = runner.getCommandInvocation("notification-configure", subReport, context.getSubject());
+            inv = runner.getCommandInvocation("requesttracing-configure", subReport, context.getSubject());
         }
 
         ParameterMap params = new ParameterMap();
         params.add("enabled", enabled.toString());
         params.add("target", target);
         params.add("dynamic", dynamic.toString());
+        params.add("thresholdUnit", unit);
+        params.add("thresholdValue", value);
         inv.parameters(params);
         inv.execute();
         // swallow the offline warning as it is not a problem
@@ -122,14 +148,14 @@ public class SetNotification implements AdminCommand {
         }
     }
 
-    private void enableNotificationNotifierConfigurerOnTarget(ActionReport actionReport, AdminCommandContext context, Boolean enabled) {
+    private void enableRequestTracingNotifierConfigurerOnTarget(ActionReport actionReport, AdminCommandContext context, Boolean enabled) {
         CommandRunner runner = serviceLocator.getService(CommandRunner.class);
         ActionReport subReport = context.getActionReport().addSubActionsReport();
 
         if (target.equals("server-config")) {
-            inv = runner.getCommandInvocation("notification-configure-notifier-das", subReport, context.getSubject());
+            inv = runner.getCommandInvocation("requesttracing-configure-notifier-das", subReport, context.getSubject());
         } else {
-            inv = runner.getCommandInvocation("notification-configure-notifier", subReport, context.getSubject());
+            inv = runner.getCommandInvocation("requesttracing-configure-notifier", subReport, context.getSubject());
         }
 
         ParameterMap params = new ParameterMap();
@@ -143,5 +169,42 @@ public class SetNotification implements AdminCommand {
         if (subReport.hasWarnings()) {
             subReport.setMessage("");
         }
+    }
+
+    private boolean validate(ActionReport actionReport) {
+        boolean result = false;
+        if (value != null) {
+            try {
+                int thresholdValue = Integer.parseInt(value);
+                if (thresholdValue <= 0 || thresholdValue > Short.MAX_VALUE * 2) {
+                    actionReport.failure(logger, "Threshold Value must be greater than zero or less than " + Short.MAX_VALUE * 2 + 1);
+                    return result;
+                }
+            } catch (NumberFormatException nfe) {
+                actionReport.failure(logger, "Threshold Value is not a valid integer", nfe);
+                return result;
+            }
+
+        }
+
+        if (unit != null) {
+            try {
+                if (!unit.equals("NANOSECONDS")
+                        && !unit.equals("MICROSECONDS")
+                        && !unit.equals("MILLISECONDS")
+                        && !unit.equals("SECONDS")
+                        && !unit.equals("MINUTES")
+                        && !unit.equals("HOURS")
+                        && !unit.equals("DAYS")) {
+                    actionReport.failure(logger, unit + " is an invalid time unit");
+                    return result;
+                }
+            } catch (IllegalArgumentException iaf) {
+                actionReport.failure(logger, unit + " is an invalid time unit", iaf);
+                return result;
+            }
+        }
+
+        return true;
     }
 }
